revert images code to master
