+++ tests/runners/airflow/liminal/liminal.yml
@@ -15,23 +15,12 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
-
 ---
 name: MyPipeline
 volumes:
   - volume: myvol1
     local:
       path: /tmp/liminal_tests
-images:
-  - image: my_python_task_img
-    type: python
-    source: write_outputs
-  - image: my_parallelized_python_task_img
-    type: python
-    source: write_outputs
-  - image: my_server_image
-    type: python
-    source: myserver
 pipelines:
   - pipeline: my_pipeline
     owner: Bosco Albert Baracus
@@ -39,7 +28,7 @@ pipelines:
     timeout_minutes: 45
     schedule: 0 * 1 * *
     default_arg_loaded: check
-    default_array_loaded: [ 2, 3, 4 ]
+    default_array_loaded: [2, 3, 4]
     default_object_loaded:
       key1: val1
       key2: val2
@@ -51,6 +40,7 @@ pipelines:
         type: python
         description: static input task
         image: my_python_task_img
+        source: write_inputs
         env_vars:
           NUM_FILES: 10
           NUM_SPLITS: 3
@@ -63,6 +53,7 @@ pipelines:
         type: python
         description: parallelized python task
         image: my_parallelized_python_task_img
+        source: write_outputs
         env_vars:
           FOO: BAR
         executors: 3
@@ -77,6 +68,7 @@ services:
     type: python_server
     description: my python server
     image: my_server_image
+    source: myserver
     endpoints:
       - endpoint: /myendpoint1
         module: my_server
