+++ tests/runners/airflow/liminal/write_outputs/write_outputs.py
@@ -16,10 +16,27 @@
 # specific language governing permissions and limitations
 # under the License.
 
-
 import json
+import logging
+import os
 
+split_id = int(os.environ['LIMINAL_SPLIT_ID'])
+num_splits = int(os.environ['LIMINAL_NUM_SPLITS'])
 
-def myendpoint1func(input_json):
-    input_dict = json.loads(input_json) if input_json else {}
-    return f'Input was: {input_dict}'
+inputs_dir = f'/mnt/vol1/inputs/{split_id}'
+outputs_dir = '/mnt/vol1/outputs/'
+
+if not os.path.exists(outputs_dir):
+    os.makedirs(outputs_dir)
+
+logging.info(f'Running write_outputs for split id {split_id} [NUM_SPLITS = {num_splits}]')
+
+for filename in os.listdir(inputs_dir):
+    with open(os.path.join(inputs_dir, filename)) as infile, \
+            open(os.path.join(
+                outputs_dir,
+                filename.replace('input', 'output').replace('.json', '.txt')
+            ), 'w') as outfile:
+        logging.info(f'Writing output file: {outfile.name}')
+        data = json.loads(infile.read())
+        outfile.write(data['mykey'])
